Run related projects migration with django migrations
I kept forgetting to run the migrations after switching to my branch and pretty much had enough of it.
Also, the strategy changed and this will be tested regularly in an environment before reaching prod,
 so we can be sure that the migration will work once we get there.

#4710: Improve project hierarchies
